decouple the logger.py and ParallelEnv
