+++ python/paddle/distributed/__init__.py
@@ -19,6 +19,7 @@ from .launch.main import launch  # noqa: F401
 from .parallel import init_parallel_env  # noqa: F401
 from .parallel import get_rank  # noqa: F401
 from .parallel import get_world_size  # noqa: F401
+from .parallel import ParallelEnv  # noqa: F401
 
 from .parallel_with_gloo import gloo_init_parallel_env
 from .parallel_with_gloo import gloo_barrier
@@ -68,8 +69,6 @@ from .fleet import BoxPSDataset  # noqa: F401
 from .entry_attr import ProbabilityEntry  # noqa: F401
 from .entry_attr import CountFilterEntry  # noqa: F401
 from .entry_attr import ShowClickEntry  # noqa: F401
-
-from ..parallel import ParallelEnv  # noqa: F401
 
 from . import cloud_utils  # noqa: F401
