+++ python/paddle/hapi/logger.py
@@ -16,11 +16,6 @@ import logging
 import os
 import sys
 
-# Here is the only case to import ParallelEnv from paddle.parallel,
-# because there is circular import if import it from paddle.distributed.
-# It is also the reason why put the implenmentation of ParallelEnv in the paddle/parallel.py
-from paddle.parallel import ParallelEnv
-
 __all__ = []
 
 
@@ -43,7 +38,7 @@ def setup_logger(output=None, name="hapi", log_level=logging.INFO):
 
     format_str = '%(asctime)s - %(name)s - %(levelname)s - %(message)s'
     # stdout logging: only local rank==0
-    local_rank = ParallelEnv().local_rank
+    local_rank = int(os.getenv("PADDLE_TRAINER_ID", "0"))
     if local_rank == 0 and len(logger.handlers) == 0:
         ch = logging.StreamHandler(stream=sys.stdout)
         ch.setLevel(log_level)
